% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_performance_measures}
\alias{get_performance_measures}
\title{get_performance_measures}
\usage{
get_performance_measures(
  score = data_BED_PLANNING_training$TOTAL_SCORE,
  truth = data_BED_PLANNING_training$ADMIT_FLAG == "Admitted",
  thresholds = 50,
  round = TRUE
)
}
\arguments{
\item{score}{(optional) A vector of scores or probabilities associated with each observation. Default: Random scores between 0 and 100.}

\item{truth}{(optional) A vector indicating the true class labels of the observations. Default: Random TRUE/FALSE values.}

\item{thresholds}{(optional) A numeric vector specifying the thresholds for binary classification. Default: 50.}

\item{round}{(optional) A logical value indicating whether to round the output metrics. Default: TRUE.}
}
\value{
A data table containing performance measures for each specified threshold, including sensitivity, specificity, balanced accuracy, accuracy, positive predictive value (ppv), negative predictive value (npv), F1 score, and Youden's index.
}
\description{
get_performance_measures description: the function calculates various performance measures for binary classification based on provided scores and true labels. It offers flexibility in specifying thresholds for binary classification and rounding options for the output metrics.
}
\examples{
# Example 1: Using default parameters
library(data.table)
performance_metrics_default <- get_performance_measures()
# Example 2: Specifying custom parameters
library(data.table)
custom_scores <- c(80, 65, 72, 90, 85)
custom_truth <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
custom_thresholds <- c(60, 70, 80)
custom_metrics <- get_performance_measures(score = custom_scores, truth = custom_truth, thresholds = custom_thresholds, round = FALSE)
}
