% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_stat}
\alias{get_stat}
\title{An Analysis Function to get statistics}
\usage{
get_stat(Actual, Prediction)
}
\arguments{
\item{Actual}{A vector of True and False}

\item{Prediction}{A vector of True and False}
}
\value{
True Positive(TP) is an outcome where the model correctly predicts the positive class.

True Negative(TN) is an outcome where the model correctly predicts the negative class.

False Positive(FP) is an outcome where the  model incorrectly predicts the positive class.

False Negative(FN) is an outcome where the model incorrectly predicts the negative class.

Sens: True positive rate, is the probability of a positive test result. conditioned on the individual truly being positive.

Spec: True negative rate, is the probability of a negative test result, conditioned on the individual truly being negative.

Accuracy: How close a given set of observations are to their true value

Balanced Accuracy: The arithmetic mean of sensitivity and specificity, often used to deal with imbalanced datasets.

Prevalence: The proportion of a population who have a specific characteristic in a given time period.

Positive Predictive Value: The proportion of positive results that are true positive.

Negative Predictive Value: The proportion of negative results that are true negative.

F1 score: The average of precision(i.e. ppv) and recall(i.e. Sens). It signifies that the model can effectively identify positive cases while minimizing false positives and false negatives.

Youden: A single statistic that captures the performance of a dichotomous diagnostic test.
}
\description{
From a Actual vector of TRUE and FALSE, and a Prediction vector of True and FALSE, calculate statistics.
}
\examples{
x = sample(c(T,F), 1000, replace = T)
y = sample(c(T,F), 1000, replace = T)
get_stat(x,y)
}
