% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_stat}
\alias{summary_stat}
\title{summary_stat}
\usage{
summary_stat(x, digits = 2, quantiles = c(5, 25, 75, 95))
}
\arguments{
\item{x}{Numeric vector for which summary statistics are computed}

\item{digits}{(optional) Number of digits to round the summary statistics to (default is 2).}

\item{quantiles}{(optional) Numeric vector specifying the quantiles to compute. Defaults to c(5, 25, 75, 95).}
}
\value{
A named numeric vector containing the computed summary statistics.
}
\description{
summary_stat description: this function computes summary statistics for a given numeric vector, including mean, standard deviation, minimum, lower quantiles, median, upper quantiles, maximum, and range.
}
\examples{
#Example 1:
data <- runif(n = 50, min = 1, max = 100)
summary_stat(data)
#Example 2:
summary_stat(data, digits = 3, quantiles = c(10, 50, 100))
}
